% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_models.R
\name{linear_function}
\alias{linear_function}
\title{Function for making the predictions based on a linear model}
\usage{
linear_function(x, my_coef)
}
\arguments{
\item{x}{named numeric vector of data}

\item{named}{numeric vector of model parameters for the linear model}
}
\description{
Function for making the predictions based on a linear model
}
